+++ src/Drogi/DrogaPowietrzna.java
@@ -15,7 +15,7 @@ public class DrogaPowietrzna extends Droga {
      * @param poczatek poczatek drogi.
      * @param koniec koniec drogi.
      */
-    public DrogaPowietrzna(MiejsceZmianyKierunku poczatek, MiejsceZmianyKierunku koniec,int poprawkaX, int poprawkaY) {
+    public DrogaPowietrzna(MiejsceZmianyKierunku poczatek, MiejsceZmianyKierunku koniec,double poprawkaX, double poprawkaY) {
         super(poczatek, koniec,poprawkaX,poprawkaY);
         this.setColor(Color.ORANGE);
         this.rysuj(MainPanel.getGrupaDrog());
