+++ src/Mapa/ZmianyKierunku/Przystanki/Miasto.java
@@ -12,7 +12,7 @@ import javafx.scene.paint.Color;
  * Created by Lewin on 2015-11-06.
  */
 public class Miasto extends Lotnisko implements TworzeniePojazdu{
-    public Miasto(String nazwa, int dlugosc, int szerokosc, int polozenieX, int polozenieY, boolean zajetaPrzestrzen, int maksymalnaPojemnosc) {
+    public Miasto(String nazwa, double dlugosc, double szerokosc, double polozenieX, double polozenieY, boolean zajetaPrzestrzen, int maksymalnaPojemnosc) {
         super(nazwa, dlugosc, szerokosc, polozenieX, polozenieY, zajetaPrzestrzen, maksymalnaPojemnosc);
         this.setColor(Color.PURPLE);
         this.rysuj(MainPanel.getGrupaMiejscZmianyKierunku());
