+++ src/Mapa/ZmianyKierunku/Przystanki/Port.java
@@ -12,7 +12,7 @@ import javafx.scene.paint.Color;
  * Created by Lewin on 2015-10-18.
  */
 public class Port extends Przystanek implements TworzeniePojazdu {
-    public Port(String nazwa, int dlugosc, int szerokosc, int polozenieX, int polozenieY, boolean zajetaPrzestrzen) {
+    public Port(String nazwa, double dlugosc, double szerokosc, double polozenieX, double polozenieY, boolean zajetaPrzestrzen) {
         super(dlugosc, szerokosc, polozenieX, polozenieY, zajetaPrzestrzen, nazwa);
         this.setColor(Color.BLUE);
         this.rysuj(MainPanel.getGrupaMiejscZmianyKierunku());
